# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for the B2B medical supplies marketplace
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products/{id}/offers:
    get:
      operationId: listProductOffers
      tags: [offers]
      summary: List vendor offers for a product, for comparison
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OfferWithVendor"
  /offers:
    post:
      operationId: createOffer
      tags: [offers]
      summary: Create a vendor offer on a master product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorOfferInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorOffer"
...
  /offers/{id}:
    patch:
      operationId: updateOffer
      tags: [offers]
      summary: Update a vendor offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorOfferUpdate"
      responses:
        "200":
          description: Updated offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorOffer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteOffer
      tags: [offers]
      summary: Delete a vendor offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /vendor/offers/bulk-upload:
    post:
      operationId: bulkUploadOffers
      tags: [offers]
      summary: Bulk-create/update vendor offers from CSV rows (productId,price,moq,stock,deliveryDays)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUploadOffersBody"
      responses:
        "200":
          description: Bulk upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkUploadResult"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    PriceTier:
      type: object
      required: [minQty, maxQty, price]
      properties:
        minQty:
          type: integer
        maxQty:
          type: [integer, "null"]
        price:
          type: number
    VendorOffer:
      type: object
      required: [id, productId, vendorCompanyId, price, moq, stock, deliveryDays, priceTiers, status, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        variantId:
          type: [integer, "null"]
        vendorCompanyId:
          type: integer
        warehouseId:
          type: [integer, "null"]
        price:
          type: number
        moq:
          type: integer
        stock:
          type: integer
        lowStockThreshold:
          type: integer
        deliveryDays:
          type: integer
        priceTiers:
          type: array
          items:
            $ref: "#/components/schemas/PriceTier"
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
    VendorOfferInput:
      type: object
      required: [productId, price, moq, stock, deliveryDays, priceTiers]
      properties:
        productId:
          type: integer
        variantId:
          type: integer
        warehouseId:
          type: integer
        lowStockThreshold:
          type: integer
          minimum: 0
        price:
          type: number
          minimum: 0
        moq:
          type: integer
          minimum: 1
        stock:
          type: integer
          minimum: 0
        deliveryDays:
          type: integer
          minimum: 0
        priceTiers:
          type: array
          items:
            $ref: "#/components/schemas/PriceTier"
    VendorOfferUpdate:
      type: object
      properties:
        price:
          type: number
        moq:
          type: integer
        stock:
          type: integer
        warehouseId:
          type: integer
        lowStockThreshold:
          type: integer
        deliveryDays:
          type: integer
        priceTiers:
          type: array
          items:
            $ref: "#/components/schemas/PriceTier"
        status:
          type: string
          enum: [active, inactive]
    OfferWithVendor:
      allOf:
        - $ref: "#/components/schemas/VendorOffer"
        - type: object
          required: [vendorName, vendorSubtype]
          properties:
            vendorName:
              type: string
            vendorSubtype:
              type: string
...
    BulkUploadOffersRow:
      type: object
      required: [productId, price, moq, stock, deliveryDays]
      properties:
        productId:
          type: integer
        price:
          type: number
        moq:
          type: integer
        stock:
          type: integer
        deliveryDays:
          type: integer
    BulkUploadOffersBody:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/BulkUploadOffersRow"
    BulkUploadResult:
      type: object
      required: [created, updated, errors]
      properties:
        created:
          type: integer
        updated:
          type: integer
        errors:
          type: array
          items:
            type: string
...