# lib/api-zod/src/generated/types/bulkUploadOffersRow.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the B2B medical supplies marketplace
 * OpenAPI spec version: 0.1.0
 */

export interface BulkUploadOffersRow {
  productId: number;
  price: number;
  moq: number;
  stock: number;
  deliveryDays: number;
}